feat: multiselection rename components - Ref gestion-de-projet#2027
