+++ src/components/ScopeTree/ScopeTreeSearch/index.tsx
@@ -1,66 +1,56 @@
 import React, { ReactElement, useEffect, useRef, useState } from 'react'
-import { ScopeTreeRow } from 'types'
-import {
-  Checkbox,
-  CircularProgress,
-  Grid,
-  Pagination,
-  Paper,
-  Table,
-  TableBody,
-  TableCell,
-  TableContainer,
-  TableHead,
-  TableRow,
-  Typography
-} from '@mui/material'
+import { Checkbox, CircularProgress, Grid, LinearProgress, Pagination } from '@mui/material'
+import { useDebounce } from 'utils/debounce'
 import EnhancedTable from '../ScopeTreeTable'
-import useStyles from '../utils/styles'
-import { AppDispatch, useAppDispatch, useAppSelector } from 'state'
 import {
   getHeadCells,
-  init,
   isSearchIndeterminate,
   isSearchSelected,
   onExpand,
-  onExplorationSelectAll,
-  onSearchSelect
+  onSearchSelect,
+  onSearchSelectAll,
+  searchInPerimeters
 } from '../utils/scopeTreeUtils'
+import useStyles from '../utils/styles'
+import { useAppSelector } from 'state'
 import { ScopeState } from 'state/scope'
-import { CareSiteExplorationProps } from '../index'
-import CareSiteHierarchy from '../CareSiteHierarchy/CareSiteHierarchy'
+import { ScopeTreeRow } from 'types'
+import { ScopeTreeSearchProps } from '../index'
+import ScopeTreeHierarchy from '../ScopeTreeHierarchy/ScopeTreeHierarchy'
 
-const Index = (props: CareSiteExplorationProps) => {
+const Index: React.FC<ScopeTreeSearchProps> = (props) => {
   const {
+    searchInput,
     selectedItems,
     setSelectedItems,
     searchRootRows,
+    setSearchRootRows,
     executiveUnitType,
     isSelectionLoading,
     setIsSelectionLoading
   } = props
 
   const { classes } = useStyles()
-  const dispatch: AppDispatch = useAppDispatch()
 
   const { scopeState } = useAppSelector<{
     scopeState: ScopeState
   }>((state) => ({
     scopeState: state.scope || {}
   }))
+
   const { scopesList = [] } = scopeState
-  const [rootRows, setRootRows] = useState<ScopeTreeRow[]>(scopesList)
-  const [isSearchLoading, setIsSearchLoading] = useState<boolean>(false)
-  const [page, setPage] = useState(1)
+  const [openPopulation, setOpenPopulations] = useState<number[]>(scopeState.openPopulation)
+  const [rootRows, setRootRows] = useState<ScopeTreeRow[]>([])
+  const controllerRef = useRef<AbortController | null>(null)
+  const [isEmpty, setIsEmpty] = useState<boolean>(true)
+  const debouncedSearchTerm = useDebounce(700, searchInput)
+  const [page, setPage] = useState(0)
   const [count, setCount] = useState(0)
-  const [isEmpty, setIsEmpty] = useState<boolean>(!rootRows || rootRows.length === 0)
-  const [openPopulation, setOpenPopulations] = useState<number[]>([])
+  const [isSearchLoading, setIsSearchLoading] = useState<boolean>(false)
 
   const isHeadChecked: boolean = rootRows.length > 0 && rootRows.every((item) => isSearchSelected(item, selectedItems))
   const isHeadIndeterminate: boolean =
     !isHeadChecked && rootRows.length > 0 && !!rootRows.find((item) => isSearchIndeterminate(item, selectedItems))
-
-  const controllerRef: React.MutableRefObject<AbortController | null> = useRef<AbortController | null>(null)
 
   const headCheckbox: ReactElement = (
     <div style={{ padding: '0 0 0 4px' }}>
@@ -69,52 +59,63 @@ const Index = (props: CareSiteExplorationProps) => {
         checked={isHeadChecked}
         indeterminate={isHeadIndeterminate}
         onClick={() =>
-          onExplorationSelectAll(rootRows, setSelectedItems, isHeadChecked, isSelectionLoading, setIsSelectionLoading)
+          onSearchSelectAll(
+            rootRows,
+            selectedItems,
+            setSelectedItems,
+            isHeadChecked,
+            searchRootRows,
+            scopesList,
+            isSelectionLoading,
+            setIsSelectionLoading
+          )
         }
       />
     </div>
   )
   const headCells = getHeadCells(headCheckbox, executiveUnitType)
 
-  useEffect(() => {
-    init(
-      setIsSearchLoading,
+  const search = async () =>
+    await searchInPerimeters(
+      debouncedSearchTerm,
+      page,
       controllerRef,
-      rootRows,
-      setRootRows,
-      setOpenPopulations,
-      setCount,
+      setIsSearchLoading,
       setIsEmpty,
-      dispatch,
+      setCount,
+      setRootRows,
+      searchRootRows,
+      setSearchRootRows,
+      setOpenPopulations,
       executiveUnitType
     )
-  }, [])
+
+  useEffect(() => {
+    if (debouncedSearchTerm) {
+      search()
+    }
+    return () => {
+      controllerRef.current?.abort()
+    }
+  }, [debouncedSearchTerm, page])
 
   return (
     <div className={classes.container}>
+      {!isSearchLoading && <div className={classes.linearProgress}>{isSelectionLoading && <LinearProgress />}</div>}
       {isSearchLoading ? (
         <Grid container justifyContent="center">
           <CircularProgress size={50} />
         </Grid>
       ) : (
         <>
-          {isEmpty ? (
-            <TableContainer component={Paper}>
-              <Table className={classes.table}>
-                <TableHead>
-                  <TableRow className={classes.tableHead}>
-                    <TableCell align="center" className={classes.tableHeadCell} />
-                  </TableRow>
-                </TableHead>
-                <TableBody>
-                  <TableRow>
-                    <TableCell colSpan={7}>
-                      <Typography className={classes.loadingSpinnerContainer}>Aucun résultat à afficher</Typography>
-                    </TableCell>
-                  </TableRow>
-                </TableBody>
-              </Table>
-            </TableContainer>
+          {!debouncedSearchTerm || isEmpty ? (
+            <EnhancedTable
+              noCheckbox
+              noPagination
+              rows={rootRows}
+              headCells={headCells}
+              emptyRowsMessage={'Aucun résultat à afficher'}
+            />
           ) : (
             <>
               <EnhancedTable noCheckbox noPagination rows={rootRows} headCells={headCells}>
@@ -123,7 +124,7 @@ const Index = (props: CareSiteExplorationProps) => {
                   const labelId = `enhanced-table-checkbox-${index}`
 
                   return (
-                    <CareSiteHierarchy
+                    <ScopeTreeHierarchy
                       row={row}
                       level={0}
                       parentAccess={row.access ?? '-'}
@@ -138,7 +139,7 @@ const Index = (props: CareSiteExplorationProps) => {
                           rootRows,
                           setRootRows,
                           selectedItems,
-                          dispatch,
+                          undefined,
                           executiveUnitType
                         )
                       }
@@ -150,11 +151,13 @@ const Index = (props: CareSiteExplorationProps) => {
                           scopesList,
                           isSelectionLoading,
                           setIsSelectionLoading,
-                          setSelectedItems
+                          setSelectedItems,
+                          setSearchRootRows
                         )
                       }
                       isIndeterminate={(row: ScopeTreeRow) => isSearchIndeterminate(row, selectedItems)}
                       isSelected={(row: ScopeTreeRow) => isSearchSelected(row, selectedItems)}
+                      isSearchMode={true}
                       executiveUnitType={executiveUnitType}
                     />
                   )
